tests/heap_cmd: use uintptr_t to store pointer addr

+++ tests/heap_cmd/main.c
@@ -42,7 +42,7 @@ static int free_cmd(int argc, char **argv)
         return 1;
     }
 
-    unsigned int p = strtoul(argv[1], NULL, 16);
+    uintptr_t p = strtoul(argv[1], NULL, 16);
     void *ptr = (void *)p;
     free(ptr);
     printf("freed %p\n", ptr);
